Move index_add to performance-tuned IndexKernel template
