+++ src/aten/sycl/Indexing.cpp
@@ -12,371 +12,20 @@
 #include <ATen/native/Resize.h>
 #include <ATen/native/TensorIterator.h>
 #include <aten/sycl/BatchKernel.h>
-#include <aten/sycl/Indexing.h>
+#include <aten/sycl/IndexingKernel.h>
 #include <aten/sycl/Loops.h>
-#include <aten/sycl/TensorInfo.h>
+
+#include <comm/TensorInfo.h>
 #include <comm/SYCLContext.h>
 #include <comm/TensorInfo.h>
-#include <stdio.h>
 
 using namespace at::xpu::detail;
 using namespace at::xpu;
 
 namespace at::native::xpu {
-namespace impl {
-
-// Pretend that the scalar tensor is in fact a one-element vector.
-template <typename T, typename IndexType>
-TensorInfo<T, IndexType> tensorInfoIfScalar(TensorInfo<T, IndexType> ti) {
-  if (ti.dims == 0) {
-    ti.dims = 1;
-    ti.sizes[0] = 1;
-    ti.strides[0] = 1;
-  }
-  return ti;
-}
-
-template <class SrcInfo, class DstInfo, class IdxInfo, class FuncType>
-class IndexKernelConfig : public BatchKernelConfig {
- public:
-  using ValType = typename SrcInfo::scalar_t;
-  using IdxType = typename IdxInfo::scalar_t;
-
-  IndexKernelConfig() = delete;
-  IndexKernelConfig(
-      SrcInfo& sinfo,
-      DstInfo& dinfo,
-      IdxInfo& iinfo,
-      ValType alpha,
-      int64_t index_num,
-      int64_t indexing_dimension_size,
-      bool indexing_dst,
-      bool problem_inner,
-      FuncType func,
-      int64_t batch,
-      int64_t problem,
-      int64_t stride,
-      int64_t problem_batch,
-      bool problem_along_x)
-      : BatchKernelConfig(
-            batch,
-            problem,
-            stride,
-            problem_batch,
-            problem_along_x,
-            Policy::pSegment,
-            syclMaxWorkItemsPerEU()),
-        sinfo_(sinfo),
-        dinfo_(dinfo),
-        iinfo_(iinfo),
-        alpha_(alpha),
-        index_num_(index_num),
-        indexing_dimension_size_(indexing_dimension_size),
-        indexing_dst_(indexing_dst),
-        problem_inner_(problem_inner),
-        func_(func) {}
-
-  template <class TarInfo>
-  static inline void indexing_problem_mapping(
-      TarInfo& tinfo,
-      IdxInfo& iinfo,
-      int dim,
-      int64_t index_num,
-      int64_t indexing_dimension_size,
-      int64_t& batch,
-      int64_t& problem,
-      int64_t& stride,
-      int64_t& problem_batch,
-      bool& problem_along_x,
-      bool& problem_inner) {
-    int64_t outer = tinfo.outerSize(dim);
-    int64_t inner = tinfo.innerSize(dim);
-
-    if (inner == 1) {
-      problem = outer;
-      stride = indexing_dimension_size;
-      batch = 1;
-      problem_batch = index_num;
-      problem_along_x = tinfo.strides[dim] == 1 ? false : true;
-      problem_inner = false;
-    } else if (outer == 1) {
-      problem = inner;
-      stride = 1;
-      batch = indexing_dimension_size;
-      problem_batch = index_num;
-      problem_along_x = tinfo.strides[tinfo.dims - 1] == 1 ? true : false;
-      problem_inner = true;
-    } else {
-      problem = inner;
-      stride = 1;
-      batch = outer * indexing_dimension_size;
-      problem_batch = outer * index_num;
-      problem_along_x = tinfo.strides[tinfo.dims - 1] == 1 ? true : false;
-      problem_inner = true;
-    }
-    return;
-  }
-
-  static IndexKernelConfig<SrcInfo, DstInfo, IdxInfo, FuncType> make_config(
-      SrcInfo& src_info,
-      DstInfo& dst_info,
-      IdxInfo& index_info,
-      ValType alpha,
-      int64_t dim,
-      bool indexing_dst,
-      FuncType func) {
-    int64_t index_num = index_info.sizes[0];
-    int64_t indexing_dimension_size;
-
-    bool problem_along_x, problem_inner;
-    int64_t batch, problem, stride, problem_batch;
-
-    TORCH_CHECK(
-        indexing_dst || src_info.data != nullptr,
-        "Indexing kernel backbone does not support null src ...");
-
-    if (indexing_dst) {
-      indexing_dimension_size = dst_info.sizes[dim];
-      indexing_problem_mapping(
-          dst_info,
-          index_info,
-          dim,
-          index_num,
-          indexing_dimension_size,
-          batch,
-          problem,
-          stride,
-          problem_batch,
-          problem_along_x,
-          problem_inner);
-    } else {
-      indexing_dimension_size = src_info.sizes[dim];
-      indexing_problem_mapping(
-          src_info,
-          index_info,
-          dim,
-          index_num,
-          indexing_dimension_size,
-          batch,
-          problem,
-          stride,
-          problem_batch,
-          problem_along_x,
-          problem_inner);
-    }
-
-    return {
-        src_info,
-        dst_info,
-        index_info,
-        alpha,
-        index_num,
-        indexing_dimension_size,
-        indexing_dst,
-        problem_inner,
-        func,
-        batch,
-        problem,
-        stride,
-        problem_batch,
-        problem_along_x};
-  }
-
- public:
-  SrcInfo sinfo_; // sinfo_.data could be nullptr, while indexing along dst.
-  DstInfo dinfo_;
-  IdxInfo iinfo_;
-  ValType alpha_;
-  int64_t index_num_;
-  int64_t indexing_dimension_size_;
-  bool indexing_dst_;
-  bool problem_inner_;
-  FuncType func_;
-};
-
-template <
-    class IdxConfig,
-    bool TrivialOffCal = false,
-    bool known_problem_inner = false>
-class IndexKernel {
- public:
-  using ValType = typename IdxConfig::ValType;
-  using IdxType = typename IdxConfig::IdxType;
-
-  IndexKernel() = delete;
-  IndexKernel(IdxConfig& cfg) : cfg_(cfg) {}
-
-  void init_global_batch_info(
-      BatchKernelConfig::ItemDesc& id,
-      int64_t& idx_logical_off,
-      int64_t& glb_batch_group,
-      int64_t& glb_batch_group_loc_off) const {
-    idx_logical_off = id.glb_batch % cfg_.index_num_;
-    int64_t idx_off;
-    if constexpr (TrivialOffCal) {
-      idx_off = idx_logical_off;
-    } else {
-      idx_off = IndexToOffset<IdxType, int64_t>::get(
-          idx_logical_off,
-          cfg_.iinfo_,
-          IndexToOffset<IdxType, int64_t>::NON_STRICT_CONTIGUOUS);
-    }
-    glb_batch_group = id.glb_batch / cfg_.index_num_;
-    glb_batch_group_loc_off = cfg_.iinfo_.data[idx_off];
-    glb_batch_group_loc_off = glb_batch_group_loc_off >= 0
-        ? glb_batch_group_loc_off
-        : cfg_.indexing_dimension_size_ + glb_batch_group_loc_off;
-  }
-
-  int64_t inline indexing_logical_off(
-      BatchKernelConfig::ItemDesc& id,
-      int64_t glb_batch_group,
-      int64_t glb_batch_group_loc_off) const {
-    int64_t si, pi, bi;
-    int64_t glb_batch_group_glb_off =
-        glb_batch_group * cfg_.indexing_dimension_size_ +
-        glb_batch_group_loc_off;
-    auto stride = cfg_.stride_;
-    if constexpr (known_problem_inner) {
-      si = 0;
-      pi = id.glb_problem;
-      bi = glb_batch_group_glb_off;
-      return (pi + bi * cfg_.problem_) * stride;
-    } else {
-      if (cfg_.problem_inner_) {
-        si = 0;
-        pi = id.glb_problem;
-        bi = glb_batch_group_glb_off;
-        return (pi + bi * cfg_.problem_) * stride;
-      } else {
-        si = glb_batch_group_glb_off;
-        pi = id.glb_problem;
-        bi = 0;
-        return si + pi * stride;
-      }
-    }
-  }
-
-  int64_t inline fixing_logical_off(
-      BatchKernelConfig::ItemDesc& id,
-      int64_t glb_batch_group,
-      int64_t idx_logical_off) const {
-    int64_t si, pi, bi, stride;
-    int64_t glb_batch_group_glb_off =
-        glb_batch_group * cfg_.index_num_ + idx_logical_off;
-    if constexpr (known_problem_inner) {
-      si = 0;
-      stride = 1;
-      pi = id.glb_problem;
-      bi = glb_batch_group_glb_off;
-      return pi + bi * cfg_.problem_;
-    } else {
-      if (cfg_.problem_inner_) {
-        si = 0;
-        stride = 1;
-        pi = id.glb_problem;
-        bi = glb_batch_group_glb_off;
-        return pi + bi * cfg_.problem_;
-      } else {
-        bi = 0;
-        si = glb_batch_group_glb_off;
-        pi = id.glb_problem;
-        stride = cfg_.index_num_;
-        return si + pi * stride;
-      }
-    }
-  }
-
-  void operator()(sycl::nd_item<2> item) const {
-    auto id = cfg_.get_item_desc(item);
-
-    if (id.glb_problem >= cfg_.problem_ ||
-        id.glb_batch >= cfg_.problem_batch_) {
-      return;
-    }
-
-    // index kernel has three operands,
-    // 1. index operand
-    // 2. operand indexing on
-    // 3. operand has fixing size as index (optional)
-    int64_t indexing_si, indexing_pi, indexing_bi;
-    int64_t fixing_si, fixing_pi, fixing_bi;
-    int64_t idx_logical_off, glb_batch_group, glb_batch_group_loc_off;
-    int64_t glb_indexing_logical_off, glb_fixing_logical_off;
-    int64_t glb_indexing_off, glb_fixing_off;
-    int64_t dst_off, src_off;
-
-    init_global_batch_info(
-        id, idx_logical_off, glb_batch_group, glb_batch_group_loc_off);
-
-    glb_indexing_logical_off =
-        indexing_logical_off(id, glb_batch_group, glb_batch_group_loc_off);
-
-    if (cfg_.sinfo_.data != nullptr && cfg_.dinfo_.data != nullptr) {
-      glb_fixing_logical_off =
-          fixing_logical_off(id, glb_batch_group, idx_logical_off);
-    }
-
-    if constexpr (TrivialOffCal) {
-      if (cfg_.indexing_dst_) {
-        dst_off = glb_indexing_logical_off;
-        if (cfg_.sinfo_.data != nullptr) {
-          src_off = glb_fixing_logical_off;
-        }
-      } else {
-        src_off = glb_indexing_logical_off;
-        dst_off = glb_fixing_logical_off;
-      }
-    } else {
-      if (cfg_.indexing_dst_) {
-        // index_copy, index_add, index_fill
-        dst_off = IndexToOffset<ValType, int64_t>::get(
-            glb_indexing_logical_off,
-            cfg_.dinfo_,
-            IndexToOffset<ValType, int64_t>::NON_STRICT_CONTIGUOUS);
-        if (cfg_.sinfo_.data != nullptr) {
-          src_off = IndexToOffset<ValType, int64_t>::get(
-              glb_fixing_logical_off,
-              cfg_.sinfo_,
-              IndexToOffset<ValType, int64_t>::NON_STRICT_CONTIGUOUS);
-        }
-      } else {
-        // index_select
-        src_off = IndexToOffset<ValType, int64_t>::get(
-            glb_indexing_logical_off,
-            cfg_.sinfo_,
-            IndexToOffset<ValType, int64_t>::NON_STRICT_CONTIGUOUS);
-        dst_off = IndexToOffset<ValType, int64_t>::get(
-            glb_fixing_logical_off,
-            cfg_.dinfo_,
-            IndexToOffset<ValType, int64_t>::NON_STRICT_CONTIGUOUS);
-      }
-    }
-    cfg_.func_(
-        cfg_.dinfo_.data,
-        cfg_.sinfo_.data,
-        dst_off,
-        src_off,
-        glb_batch_group_loc_off,
-        cfg_.alpha_);
-  }
-
- private:
-  IdxConfig cfg_;
-};
-
-template <
-    class IdxConfig,
-    bool TrivialOffCal = false,
-    bool known_problem_inner = false>
-static inline void launch_index_kernel(IdxConfig& cfg) {
-  auto& queue = getCurrentSYCLQueue();
-  IndexKernel<IdxConfig, TrivialOffCal, known_problem_inner> idx_ker(cfg);
-  sycl_kernel_submit(cfg.global_size(), cfg.group_size(), queue, idx_ker);
-}
 
 template <typename ValType>
-class IndexSelectOperator {
+class IndexSelectScalarFunctor {
  public:
   void operator()(
       ValType* dst,
@@ -404,7 +53,7 @@ static inline void _index_select_kernel(
       SrcInfo,
       DstInfo,
       IdxInfo,
-      IndexSelectOperator<scalar_t>>::
+      IndexSelectScalarFunctor<scalar_t>>::
       make_config(
           src_info,
           dst_info,
@@ -412,7 +61,7 @@ static inline void _index_select_kernel(
           static_cast<scalar_t>(0),
           dim,
           false,
-          IndexSelectOperator<scalar_t>());
+          IndexSelectScalarFunctor<scalar_t>());
   if (cfg.problem_inner_) {
     launch_index_kernel<decltype(cfg), TrivialOffCal, true>(cfg);
   } else {
@@ -421,7 +70,7 @@ static inline void _index_select_kernel(
 }
 
 template <typename scalar_t>
-void IndexSelect(
+void index_select_kernel(
     const Tensor& dst,
     const Tensor& src,
     int dim,
@@ -486,6 +135,8 @@ void IndexSelect(
         tensorInfoIfScalar(getTensorInfo<scalar_t, int64_t>(src.contiguous()));
     int new_indexing_dim = src_info.collapseDims(dim);
 
+    // Improve efficiency of generated native instructions for contiguous.
+    // See comm/TensorInfo.h
     if (dst.is_contiguous() && indices.is_contiguous())
       _index_select_kernel<
           decltype(src_info),
@@ -503,28 +154,6 @@ void IndexSelect(
   });
   return;
 }
-
-} // namespace impl
-
-Tensor& index_select_out_kernel(
-    const Tensor& self,
-    int64_t dim,
-    const Tensor& index,
-    Tensor& out) {
-  AT_DISPATCH_V2(
-      self.scalar_type(),
-      "index_select",
-      AT_WRAP([=]() { impl::IndexSelect<scalar_t>(out, self, dim, index); }),
-      AT_EXPAND(AT_ALL_TYPES_AND_COMPLEX),
-      AT_EXPAND(AT_BAREBONES_UNSIGNED_TYPES),
-      at::ScalarType::ComplexHalf,
-      at::ScalarType::Half,
-      at::ScalarType::BFloat16,
-      at::ScalarType::Bool);
-  return out;
-}
-
-using namespace at::xpu::detail;
 
 template <typename scalar_t>
 struct MaskedFillFunctor {
@@ -608,218 +237,21 @@ static ptrdiff_t getSliceSize(
   return dstSliceSize;
 }
 
-// We prefer this kernel to avoid reloading index points if the number
-// of indices is a small number.
-// This kernel in fact works for all choices of problem size, but if
-// the number of indices chosen is large, then the
-// indexFuncLargeIndex kernel is a better choice to increase
-// parallelism.
-template <
-    typename T,
-    typename IndicesType,
-    typename IndexType,
-    int DstDim,
-    int SrcDim,
-    int IdxDim,
-    typename func_t>
-struct IndexFuncSmallIndexFunctor {
-  void operator()(sycl::nd_item<1> item) const {
-    // In order to avoid reloading the index that we are copying, load
-    // it once to handle all of the points that are being selected, so
-    // it can be reused as much as possible. This kernel is chosen when
-    // this is a good choice (small number of chosen indices), since
-    // re-accessing indices in addition to src elements can be slow.
-    for (IndexType srcIndex = 0; srcIndex < indices_.sizes[0]; ++srcIndex) {
-      // Lua indices begin at 1
-      IndexType dstIndex =
-          indices_
-              .data[IndexToOffset<const IndicesType, IndexType, IdxDim>::get(
-                  srcIndex, indices_)];
-
-      // We stride over the output ignoring the indexed dimension
-      // (innerSize), whose offset calculation is handled differently
-      for (IndexType linearIndex = item.get_global_linear_id();
-           linearIndex < innerSize_;
-           linearIndex += item.get_group_range(0) * item.get_local_range(0)) {
-        IndexType dstOffset =
-            IndexToOffset<T, IndexType, DstDim>::get(linearIndex, dst_);
-        dstOffset += dstIndex * dst_.strides[dstAddDim_];
-
-        IndexType srcOffset =
-            IndexToOffset<const T, IndexType, SrcDim>::get(linearIndex, src_);
-        srcOffset += srcIndex * src_.strides[srcAddDim_];
-
-        T val;
-        if constexpr (std::is_same<T, bool>::value) {
-          val = src_.data[srcOffset] && alpha_;
-        } else {
-          val = src_.data[srcOffset] * alpha_;
-        }
-        op_(dst_.data, dstOffset, dstNumel_, &val);
-      }
-    }
-  }
-
-  IndexFuncSmallIndexFunctor(
-      TensorInfo<T, IndexType> dst,
-      TensorInfo<const T, IndexType> src,
-      TensorInfo<const IndicesType, IndexType> indices,
-      int dstAddDim,
-      int srcAddDim,
-      IndexType innerSize,
-      int64_t dstAddDimSize,
-      int64_t dstNumel,
-      func_t op,
-      T alpha)
-      : dst_(dst),
-        src_(src),
-        indices_(indices),
-        dstAddDim_(dstAddDim),
-        srcAddDim_(srcAddDim),
-        innerSize_(innerSize),
-        dstAddDimSize_(dstAddDimSize),
-        dstNumel_(dstNumel),
-        op_(op),
-        alpha_(alpha) {}
-
- private:
-  TensorInfo<T, IndexType> dst_;
-  TensorInfo<const T, IndexType> src_;
-  TensorInfo<const IndicesType, IndexType> indices_;
-  int dstAddDim_;
-  int srcAddDim_;
-  IndexType innerSize_;
-  int64_t dstAddDimSize_;
-  int64_t dstNumel_;
-  func_t op_;
-  T alpha_;
-};
-
-// We prefer this kernel to balance parallelism across index points,
-// if there are a large number of indices.
-// This kernel in fact works for all choices of problem size, but if
-// the number of indices chosen is small, then the
-// indexFuncSmallIndex kernel is a better choice to reduce memory
-// accesses.
-template <
-    typename T,
-    typename IndicesType,
-    typename IndexType,
-    int DstDim,
-    int SrcDim,
-    int IdxDim,
-    bool IndexIsMajor,
-    typename func_t>
-struct IndexFuncLargeIndexFunctor {
-  void operator()(sycl::nd_item<1> item) const {
-    // We stride over the output including the indexed dimension
-    // (totalSize), and calculate the destination index point based on that
-    for (IndexType linearIndex = item.get_global_linear_id();
-         linearIndex < totalSize_;
-         linearIndex += item.get_group_range(0) * item.get_local_range(0)) {
-      IndexType srcIndex, elementInSlice;
-      if (IndexIsMajor) {
-        srcIndex = linearIndex / innerSize_;
-        elementInSlice = linearIndex % innerSize_;
-      } else {
-        elementInSlice = linearIndex / innerSize_;
-        srcIndex = linearIndex % innerSize_;
-      }
-
-      // Lua indices begin at 1
-      IndexType dstIndex =
-          indices_
-              .data[IndexToOffset<const IndicesType, IndexType, IdxDim>::get(
-                  srcIndex, indices_)];
-
-      IndexType dstOffset =
-          IndexToOffset<T, IndexType, DstDim>::get(elementInSlice, dst_);
-      dstOffset += dstIndex * dst_.strides[dstAddDim_];
-
-      IndexType srcOffset =
-          IndexToOffset<const T, IndexType, SrcDim>::get(elementInSlice, src_);
-      srcOffset += srcIndex * src_.strides[srcAddDim_];
-
-      T val;
-      if constexpr (std::is_same<T, bool>::value) {
-        val = src_.data[srcOffset] && alpha_;
-      } else {
-        val = src_.data[srcOffset] * alpha_;
-      }
-      op_(dst_.data, dstOffset, dstNumel_, &val);
-    }
-  }
-
-  IndexFuncLargeIndexFunctor(
-      TensorInfo<T, IndexType> dst,
-      TensorInfo<const T, IndexType> src,
-      TensorInfo<const IndicesType, IndexType> indices,
-      int dstAddDim,
-      int srcAddDim,
-      IndexType totalSize,
-      IndexType innerSize,
-      int64_t dstAddDimSize,
-      int64_t dstNumel,
-      func_t op,
-      T alpha)
-      : dst_(dst),
-        src_(src),
-        indices_(indices),
-        dstAddDim_(dstAddDim),
-        srcAddDim_(srcAddDim),
-        totalSize_(totalSize),
-        innerSize_(innerSize),
-        dstAddDimSize_(dstAddDimSize),
-        dstNumel_(dstNumel),
-        op_(op),
-        alpha_(alpha) {}
-
- private:
-  TensorInfo<T, IndexType> dst_;
-  TensorInfo<const T, IndexType> src_;
-  TensorInfo<const IndicesType, IndexType> indices_;
-  int dstAddDim_;
-  int srcAddDim_;
-  IndexType totalSize_;
-  IndexType innerSize_;
-  int64_t dstAddDimSize_;
-  int64_t dstNumel_;
-  func_t op_;
-  T alpha_;
-};
-
-template <typename scalar_t>
-bool indexShouldBeMajor(
-    TensorInfo<scalar_t, unsigned int>& info,
-    int sliceDim) {
-  // The stride between adjacent slices (e.g., between element #0 of slice #100
-  // and element #0 of slice #101).
-  unsigned int sliceStride = info.strides[sliceDim];
-
-  for (const auto i : c10::irange(info.dims)) {
-    if (i != sliceDim && info.sizes[i] > 1 && info.strides[i] < sliceStride) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-template <typename scalar_t>
-struct ReduceAdd {
-  inline void operator()(
-      scalar_t* self_data_start,
-      int64_t index,
-      int64_t numel,
-      const scalar_t* src_data) const {
-    // TODO: enable fast atomic add
-    sycl_global_ptr<scalar_t> out_ptr = {self_data_start + index};
-    auto in = *src_data;
-    atomicAdd(out_ptr, in);
+template <typename ValType>
+struct IndexAddScalarFunctor {
+  void operator()(
+      ValType* dst,
+      ValType* src,
+      int64_t dst_off,
+      int64_t src_off,
+      int64_t idx,
+      ValType alpha) const {
+    atomicAdd(
+        (sycl_global_ptr<ValType>)(dst + dst_off), src[src_off] * alpha);
   }
 };
 
-void index_add_impl(
+void index_add_kernel(
     const Tensor& self,
     int64_t dim,
     const Tensor& index,
@@ -869,179 +301,33 @@ void index_add_impl(
   const ptrdiff_t sliceSize = getSliceSize(self_, dim, index, source_);
   const ptrdiff_t sourceTotalSize = source.numel();
   const int64_t selfAddDimSize = self_.size(dim);
-  const ptrdiff_t numIndex = index.numel();
-  const int64_t selfNumel = self_.numel();
 
   if (sliceSize == 0) {
     return;
   }
 
-  bool indContig = index.is_contiguous();
-  const int ssc = syclMaxDSSNum();
+  TensorInfo<int64_t, int64_t> indices_info =
+      getTensorInfo<int64_t, int64_t>(indices);
+  indices_info.collapseDims();
 
-#define SMALL_INDEX(                                                \
-    TENSOR_TYPE, INDICES_TYPE, TYPE, SELF_DIM, SOURCE_DIM, IDX_DIM) \
-  {                                                                 \
-    auto caller = IndexFuncSmallIndexFunctor<                       \
-        TENSOR_TYPE,                                                \
-        INDICES_TYPE,                                               \
-        TYPE,                                                       \
-        SELF_DIM,                                                   \
-        SOURCE_DIM,                                                 \
-        IDX_DIM,                                                    \
-        ReduceAdd<scalar_t>>(                                       \
-        selfInfo,                                                   \
-        sourceInfo,                                                 \
-        indexInfo,                                                  \
-        selfAddDim,                                                 \
-        sourceAddDim,                                               \
-        sliceSize,                                                  \
-        selfAddDimSize,                                             \
-        selfNumel,                                                  \
-        ReduceAdd<scalar_t>(),                                      \
-        alpha_value);                                               \
-    sycl_kernel_submit(                                             \
-        small_index_num_groups* small_index_group_size,             \
-        small_index_group_size,                                     \
-        getCurrentSYCLQueue(),                                      \
-        caller);                                                    \
-  }
+  TensorInfo<scalar_t, int64_t> src_info =
+      getTensorInfo<scalar_t, int64_t>(source_);
 
-#define LARGE_INDEX(                                    \
-    TENSOR_TYPE,                                        \
-    INDICES_TYPE,                                       \
-    TYPE,                                               \
-    SELF_DIM,                                           \
-    SOURCE_DIM,                                         \
-    IDX_DIM,                                            \
-    IDX_IS_MAJOR)                                       \
-  {                                                     \
-    auto caller = IndexFuncLargeIndexFunctor<           \
-        TENSOR_TYPE,                                    \
-        INDICES_TYPE,                                   \
-        TYPE,                                           \
-        SELF_DIM,                                       \
-        SOURCE_DIM,                                     \
-        IDX_DIM,                                        \
-        IDX_IS_MAJOR,                                   \
-        ReduceAdd<scalar_t>>(                           \
-        selfInfo,                                       \
-        sourceInfo,                                     \
-        indexInfo,                                      \
-        selfAddDim,                                     \
-        sourceAddDim,                                   \
-        sourceTotalSize,                                \
-        (IDX_IS_MAJOR) ? sliceSize : numIndex,          \
-        selfAddDimSize,                                 \
-        selfNumel,                                      \
-        ReduceAdd<scalar_t>(),                          \
-        alpha_value);                                   \
-    sycl_kernel_submit(                                 \
-        large_index_num_groups* large_index_group_size, \
-        large_index_group_size,                         \
-        getCurrentSYCLQueue(),                          \
-        caller);                                        \
-  }
+  TensorInfo<scalar_t, int64_t> dst_info =
+      getTensorInfo<scalar_t, int64_t>(self_);
+  int new_indexing_dim = dst_info.collapseDims(dim);
 
-  auto small_index_num_groups =
-      std::min(ceil_div(sliceSize, (ptrdiff_t)256), (ptrdiff_t)(ssc * 8));
-  auto small_index_group_size = std::min(sliceSize, (ptrdiff_t)256);
-  auto large_index_num_groups =
-      std::min(ceil_div(sourceTotalSize, (ptrdiff_t)256), (ptrdiff_t)(ssc * 8));
-  auto large_index_group_size = std::min(sourceTotalSize, (ptrdiff_t)256);
-
-  if (canUse32BitIndexMath(result) && canUse32BitIndexMath(source) &&
-      canUse32BitIndexMath(index)) {
-    AT_DISPATCH_ALL_TYPES_AND_COMPLEX_AND4(
-        at::ScalarType::Bool,
-        at::ScalarType::Half,
-        at::ScalarType::BFloat16,
-        at::ScalarType::ComplexHalf,
-        result.scalar_type(),
-        "index_add",
-        [&] {
-          TensorInfo<scalar_t, unsigned int> selfInfo =
-              getTensorInfo<scalar_t, unsigned int>(self_);
-          const int selfAddDim = selfInfo.collapseDims(dim);
-          selfInfo.reduceDim(selfAddDim);
-          const auto alpha_value = alpha.to<scalar_t>();
-          AT_DISPATCH_INDEX_TYPES(index.scalar_type(), "index_add_xpu_", [&]() {
-            auto sourceInfo =
-                getTensorInfo<const scalar_t, unsigned int>(source_);
-            const int sourceAddDim = sourceInfo.collapseDims(dim);
-            sourceInfo.reduceDim(sourceAddDim);
-
-            auto indexInfo = getTensorInfo<const index_t, unsigned int>(index);
-            indexInfo.collapseDims();
-
-            // A reasonable choice for when to have each thread iterate over
-            // index to choose
-            if (numIndex <= 16) {
-              if (selfInfo.dims == 1 && sourceInfo.dims == 1 && indContig) {
-                SMALL_INDEX(scalar_t, index_t, unsigned int, 1, 1, -2);
-              } else if (
-                  selfInfo.dims == 2 && sourceInfo.dims == 2 && indContig) {
-                SMALL_INDEX(scalar_t, index_t, unsigned int, 2, 2, -2);
-              } else if (
-                  selfInfo.dims == 3 && sourceInfo.dims == 3 && indContig) {
-                SMALL_INDEX(scalar_t, index_t, unsigned int, 3, 3, -2);
-              } else {
-                SMALL_INDEX(scalar_t, index_t, unsigned int, -1, -1, -1);
-              }
-            } else {
-              const bool indexIsMajor =
-                  indexShouldBeMajor(selfInfo, selfAddDim);
-
-              if (selfInfo.dims == 1 && sourceInfo.dims == 1 && indContig) {
-                LARGE_INDEX(scalar_t, index_t, unsigned int, 1, 1, -2, true);
-              } else if (
-                  selfInfo.dims == 2 && sourceInfo.dims == 2 && indContig) {
-                if (indexIsMajor) {
-                  LARGE_INDEX(scalar_t, index_t, unsigned int, 2, 2, -2, true);
-                } else {
-                  LARGE_INDEX(scalar_t, index_t, unsigned int, 2, 2, -2, false);
-                }
-              } else if (
-                  selfInfo.dims == 3 && sourceInfo.dims == 3 && indContig) {
-                if (indexIsMajor) {
-                  LARGE_INDEX(scalar_t, index_t, unsigned int, 3, 3, -2, true);
-                } else {
-                  LARGE_INDEX(scalar_t, index_t, unsigned int, 3, 3, -2, false);
-                }
-              } else {
-                LARGE_INDEX(scalar_t, index_t, unsigned int, -1, -1, -1, true);
-              }
-            }
-          });
-        });
-  } else {
-    AT_DISPATCH_ALL_TYPES_AND_COMPLEX_AND3(
-        at::ScalarType::Bool,
-        at::ScalarType::Half,
-        at::ScalarType::BFloat16,
-        self.scalar_type(),
-        "index_add",
-        [&] {
-          TensorInfo<scalar_t, uint64_t> selfInfo =
-              getTensorInfo<scalar_t, uint64_t>(self_);
-          const int selfAddDim = selfInfo.collapseDims(dim);
-          selfInfo.reduceDim(selfAddDim);
-          const auto alpha_value = alpha.to<scalar_t>();
-
-          TensorInfo<const scalar_t, uint64_t> sourceInfo =
-              getTensorInfo<const scalar_t, uint64_t>(source_);
-          const int sourceAddDim = sourceInfo.collapseDims(dim);
-          sourceInfo.reduceDim(sourceAddDim);
-
-          AT_DISPATCH_INDEX_TYPES(index.scalar_type(), "index_add_xpu_", [&]() {
-            TensorInfo<const index_t, uint64_t> indexInfo =
-                getTensorInfo<const index_t, uint64_t>(index);
-            indexInfo.collapseDims();
-
-            LARGE_INDEX(scalar_t, index_t, uint64_t, -1, -1, -1, true);
-          });
-        });
-  }
+  auto cfg =
+      IndexKernelConfig<decltype(src_info), decltype(dst_info), decltype(indices_info), IndexAddScalarFunctor<scalar_t>>::
+          make_config(
+              src_info,
+              dst_info,
+              index_info,
+              alpha,
+              dim,
+              true,
+              IndexAddScalarFunctor<scalar_t>());
+  launch_index_kernel(cfg);
 
 #undef SMALL_INDEX
 #undef LARGE_INDEX
